backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
shap
requests
python-whois
bs4
